weather-app/app/requirements.txt
pip~=22.0.2
wheel~=0.37.1
Flask~=3.0.0
requests~=2.25.1
python-dotenv~=1.0.0
gunicorn
